4/9
fixed the navigation bar - When you hover over tabs that take you to a different page, I changed it to a dark purple (that also passed the accessibility test). I also edited the code so all 4 main tabs stay at the top when the mediums dropdown option happens. The mediums dropdown has rounded edges and goes overtop my name (it took a minute to figure out why my name showed through). As the mediums tab doesn't take you to a new page, I instead added an underline option on the hover effect so it still has user interface.

I decided to center my name and keep the year I was born off-set. I think it makes the page look neater when the size of the screen decreases.

The photo took a while to figure out the padding and how to get the text to wrap it.

For the paragraphs, I originally used breaks (<br><br>) - but I then decided to make each paragraph its own and assign them different classes. I adjusted the margin-top/bottom sizes so there will be a gap between paragraphs. I also added indents so it looks a bit cleaner.

For the footer, I ended up adding a divider between the SVGs/icons as I didn't want them so close together or too far apart.  I also embeded the respective link to each SVG. I'm seeing if I can add a hover option where a border appears around it (I have tried, but it looked off). For now, I like that the mouse changes to a hand/white glove when you hover the SVGs/icons. I origonally coded them as buttons, but then I changed it to <a><img=""></a> and saw how it changed the mouse.

On all pages, I added a <h1> that has top/bottom margins added in CSS.