4/9
Updated the biography photo so it was more clear. I also realized that the image itself had space above and below (which is why I was struggling with the formatting for a while), so I got rid of that.

I added my animations to the animations page and gave them descriptions.